chore: move TestCosmos into tests folder.
